Split the list-item edit modal into a shell and an on-demand form
Admin::EditListItemModalComponent now renders one argument-free <dialog>
shell per page (mirroring the wizard's SharedModalComponent), with a
turbo_frame_tag placeholder. The old per-item form body moves to a new
Admin::EditListItemFormComponent, rendered into that frame, keeping the
DIALOG_ID/FRAME_ID handshake modal_form_controller.js relies on to close
the dialog after a successful save.

Wiring the view to load the form on demand is Task 2; the view still
calls the old EditListItemModalComponent.new(list_item:) signature, so
19 controller-test failures in Admin::ListItemsControllerTest and
Admin::DomainIsolationTest are expected until that lands.
