+++ web-app/app/components/admin/edit_list_item_form_component.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+class Admin::EditListItemFormComponent < ViewComponent::Base
+  def initialize(list_item:)
+    @list_item = list_item
+    @list = list_item.list
+    @config = Admin::DomainRouting.list_config(@list) || {}
+  end
+
+  def autocomplete_url
+    @config[:autocomplete_path]
+  end
+
+  def item_label
+    @config.fetch(:item_label, "Item")
+  end
+
+  def item_display_name
+    return unverified_item_display_name if @list_item.listable.nil?
+
+    if @list_item.listable.respond_to?(:title)
+      @list_item.listable.title
+    elsif @list_item.listable.respond_to?(:name)
+      @list_item.listable.name
+    else
+      "#{@list_item.listable.class.name} ##{@list_item.listable.id}"
+    end
+  end
+
+  def unverified_item_display_name
+    if @list_item.metadata.present?
+      @list_item.metadata["title"] || @list_item.metadata["name"] || "Unverified Item ##{@list_item.position}"
+    else
+      "Unverified Item ##{@list_item.position}"
+    end
+  end
+
+  def metadata_json
+    return "" if @list_item.metadata.blank?
+    JSON.pretty_generate(@list_item.metadata)
+  end
+end
